delete workaround (avoid errors when client is sending wrong deletes)

+++ src/jtl/Connector/Modified/Mapper/Category.php
@@ -113,11 +113,15 @@ class Category extends \jtl\Connector\Modified\Mapper\BaseMapper
 
     public function delete($data)
     {
-        $this->db->query('DELETE FROM categories WHERE categories_id='.$data->getId()->getEndpoint());
-        $this->db->query('DELETE FROM categories_description WHERE categories_id='.$data->getId()->getEndpoint());
-        $this->db->query('DELETE FROM products_to_categories WHERE categories_id='.$data->getId()->getEndpoint());
+        $id = $data->getId()->getEndpoint();
 
-        $this->db->query('DELETE FROM jtl_connector_link WHERE type=1 && endpointId="'.$data->getId()->getEndpoint().'"');
+        if (!empty($id) && $id != '') {
+            $this->db->query('DELETE FROM categories WHERE categories_id='.$data->getId()->getEndpoint());
+            $this->db->query('DELETE FROM categories_description WHERE categories_id='.$data->getId()->getEndpoint());
+            $this->db->query('DELETE FROM products_to_categories WHERE categories_id='.$data->getId()->getEndpoint());
+
+            $this->db->query('DELETE FROM jtl_connector_link WHERE type=1 && endpointId="'.$data->getId()->getEndpoint().'"');
+        }
 
         return $data;
     }
